test(rimap-server): add in-process TestDaemon harness
Extract `build_registry` from `main.rs` into the public
`boot::registry::build` async function so both the daemon and integration
tests share one code path.  Add `tests/common/daemon_harness.rs` with
`TestDaemon::spawn` (boots a daemon accept loop in a tokio task against a
tempdir) and `TestDaemon::shutdown` (signals + awaits the task).  The
`{tempdir}` placeholder in config strings is substituted at spawn time so
callers can reference the tempdir in `audit.path` / `allowed_base_dir`
without knowing the path ahead of time.
