+++ crates/rimap-server/tests/common/mod.rs
@@ -1,0 +1,1 @@
+pub mod daemon_harness;
